(test) Moved getTestingToken to spec file
